Split template string tokenizing into split() method.
Add `Lexer.split()` and `DebugLexer.split()` to handle the actual
splitting of tokens with a standard return value as a 2-tuple of
`(substring, position)`. `Lexer.split()` always returns `None` as a
value for `position`.

This means that `DebugLexer` no longer needs to override `.tokenize()`
from `Lexer` and they can share a common implementation.