Refs #????? -- Added specifier constants to `django.utils.dateformat.Formatter`.
This allows us to remove the need for checking whether a specifier
method is on the `TimeFormat` subclass so that we can merge all methods
into the one `Formatter` class. Containment check on a set is also much
faster than `isinstance()`.

Note that this change has added the `e`, `O`, `T`, and `Z` specifiers
into the `test_invalid_time_format_specifiers` test where they were not
previously included, but based on the code it looks as though they
should have been present. The test passes as expected.